+++ ModDebuggingMod/Card.cs
@@ -10,8 +10,8 @@ namespace ModDebuggingMod
         private void CARD_DEBUG()
         {
             CardInfo info = CardHelper.NewCard(true, "wstl", "wstlcard", "Debug",
-                attack: 1, health: 100, blood: 0, bones: 0, energy: 0, gems: null)
-                .AddAbilities(Ability.Reach, Test.ability)
+                attack: 31, health: 100, blood: 0, bones: 0, energy: 0, gems: null)
+                .AddAbilities(Ability.Reach)
                 //.AddSpecialAbilities(Bind.specialAbility)
                 //.SetTransformerCardId("Squirrel")//.SetEvolve("Squirrel", 1)
                 .SetPortraits("misterWin_grimora", emissionName: "misterWin_grimora_emission", pixelPortraitName: "buffBell.png")