+++ WhistleWind.AbnormalSigils/Abilities/Status Effects/Haste.cs
@@ -39,17 +39,15 @@ namespace WhistleWind.AbnormalSigils
         {
             // LightNegationEffect sets DoingAttackAnimation to false for some reason, which causes visual glitches
             // so do it this way
+            EffectSeverity = 0;
             base.PlayableCard.Anim.NegationEffect(false);
             CardModificationInfo mod = base.PlayableCard.TemporaryMods.Find(x => x.singletonId == CardModSingletonName);
             if (mod != null)
                 base.PlayableCard.RemoveTemporaryMod(mod);
 
-            // remove this trigger from the card
-            base.PlayableCard.TriggerHandler.specialAbilities.RemoveAll(x => x.Item1 == specialAbility);
-            Destroy();
-
             yield return new WaitForSeconds(0.25f);
         }
+
         public int TriggerPriority(bool playerIsAttacker, List<CardSlot> originalSlots) => 0;
     }
     public partial class AbnormalPlugin